Fix issue where some comments went missing
Problem was a "SocketException: Socket closed" after trying to read the
response of posting a comment to Atelier. This exception would then
raise through and prevent sending all other comments too. The reason for
this socket closure was a line in the generic API method that released
the request immediately after getting a response. For most small
responses we could still read the content before the socket would
actually be closed and the data discarded, but for some large files this
would take too long. The resulting failures turned out to be
surprisingly deterministic for such a timing issue, but removing the
call to release and switching to closeable responses (which should be
closed in the client code :( ) does seem to solve the issue.